refact: Split entity to entity package
