+++ src/main/java/gift/repository/MemberJpaDao.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.dto.Member;
+import gift.entity.Member;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
