+++ src/main/java/gift/repository/ProductJpaDao.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.dto.Product;
+import gift.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductJpaDao extends JpaRepository<Product, Long> {
