+++ src/main/java/gift/repository/WishlistJpaDao.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.dto.Wishlist;
+import gift.entity.Wishlist;
 import jakarta.transaction.Transactional;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
