+++ src/test/java/gift/repository/MemberJpaDaoTest.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.dto.Member;
+import gift.entity.Member;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
